Removed premature clearance of password for AD
In case you have an M$ AD as LDAP-Server and configured it as elytron realm, you have to use the `DirectEvidenceVerifier`.

Once you have users coming from this Realm, no "HTTP-Remoting" calls are possible, 'cuz the password is set to `0x00` within the verifying. So when the system tries to connect to a second system using the stored credentials, it fails.

We tracked down the problem only for AD-User. A parallel JDBC-Realm can do the remote calls without problems. Through debugging we found out, that the `DirectEvidenceVerifier` clears (set to series of `0x00`) to early (premature), so the caller stores the wrong credentials (`0x00`) in heap for later remote calls after verifying it.